// src/classes/Entity/types.d.ts
export interface EntityConstructor<EntityTable extends TableDef | undefined = undefined, Name extends string = string, AutoExecute extends boolean = true, AutoParse extends boolean = true, Timestamps extends boolean = true, CreatedAlias extends string = 'created', ModifiedAlias extends string = 'modified', TypeAlias extends string = 'entity', TypeHidden extends boolean = false, ReadonlyAttributeDefinitions extends Readonly<AttributeDefinitions> = Readonly<AttributeDefinitions>> {
    table?: EntityTable;
    name: Name;
    timestamps?: Timestamps;
    created?: string;
    modified?: string;
    createdAlias?: CreatedAlias;
    modifiedAlias?: ModifiedAlias;
    typeAlias?: TypeAlias;
    typeHidden?: TypeHidden;
    attributes: ReadonlyAttributeDefinitions;
    autoExecute?: AutoExecute;
    autoParse?: AutoParse;
}
// ...
export declare type PartitionKeyDefinition = Partial<KeyAttributeDefinition> & {
    partitionKey: true;
    sortKey?: false;
};
export declare type GSIPartitionKeyDefinition = Partial<KeyAttributeDefinition> & {
    partitionKey: string;
    sortKey?: false;
};
export declare type SortKeyDefinition = Partial<KeyAttributeDefinition> & {
    sortKey: true;
    partitionKey?: false;
};
export declare type GSISortKeyDefinition = Partial<KeyAttributeDefinition> & {
    partitionKey?: false;
    sortKey: string;
};
export declare type PureAttributeDefinition = Partial<{
    partitionKey: false;
    sortKey: false;
    type: DynamoDBTypes;
    default: any | ((data: object) => any);
    dependsOn: string | string[];
    transform: (value: any, data: {}) => any;
    format: (value: any, data: {}) => any;
    coerce: boolean;
    save: boolean;
    onUpdate: boolean;
    hidden: boolean;
    required: boolean | 'always';
    alias: string;
    map: string;
    setType: DynamoDBKeyTypes;
    delimiter: string;
    prefix: string;
    suffix: string;
}>;
export declare type CompositeAttributeDefinition = [string, number] | [string, number, DynamoDBTypes] | [string, number, PureAttributeDefinition];
export declare type AttributeDefinition = DynamoDBTypes | PartitionKeyDefinition | SortKeyDefinition | GSIPartitionKeyDefinition | GSISortKeyDefinition | PureAttributeDefinition | CompositeAttributeDefinition;
export declare type AttributeDefinitions = Record<A.Key, AttributeDefinition>;
// ...
export declare type Readonly<T> = T extends F.Function | undefined ? T : T extends O.Object ? {
    readonly [P in keyof T]: Readonly<T[P]>;
} : T;